homework fragments add main screen part 3
