+++ app/src/main/AndroidManifest.xml
@@ -14,6 +14,9 @@
         android:theme="@style/Theme.Fragments"
         tools:targetApi="31">
         <activity
+            android:name=".Task2Activity"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
@@ -22,6 +25,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".Task1Activity"
+            android:exported="false" />
     </application>
 
 </manifest>
