+++ app/src/main/java/otus/gpb/homework/fragments/Task2Activity.kt
@@ -1,0 +1,11 @@
+package otus.gpb.homework.fragments
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class Task2Activity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_task_2)
+    }
+}
