+++ app/src/main/res/layout/activity_task_1.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".Task1Activity">
+
+    <Button
+        style="@style/Widget.Material3.Button.Icon"
+        android:id="@+id/btnOpenFragmentA"
+        android:textSize="20sp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:text="@string/open_fragment_a" />
+
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/frgmtAContainer"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
+
+</FrameLayout>
